AdminAction changes
Since several methods in MenuAction.cs are not static, it became necessary to create objects of MenuAction to call on these methods.

In AdminActions, the user/admin would not be able to exit the menu because the previous "to-main-menu" method called became repurposed and another method had to be used, which also needed an object since it wasn't static. Issue is solved